+++ packages/transaction-controller/src/types.ts
@@ -292,6 +292,9 @@ export type TransactionMeta = {
    */
   originalType?: TransactionType;
 
+  /** Metadata specific to the MetaMask Pay feature. */
+  metamaskPay?: MetamaskPayMetadata;
+
   /**
    * Account transaction balance after swap.
    */
@@ -341,6 +344,12 @@ export type TransactionMeta = {
    * When the transaction is dropped, this is the replacement transaction ID.
    */
   replacedById?: string;
+
+  /**
+   * IDs of any transactions that must be confirmed before this one is submitted.
+   * Unlike a transaction batch, these transactions can be on alternate chains.
+   */
+  requiredTransactionIds?: string[];
 
   /**
    * The number of times that the transaction submit has been retried.
@@ -1909,4 +1918,22 @@ export type AssetsFiatValues = {
    * The fiat value of the sending assets.
    */
   sending?: string;
+};
+
+/** Metadata specific to the MetaMask Pay feature. */
+export type MetamaskPayMetadata = {
+  /** Total fee from any bridge transactions, in fiat currency. */
+  bridgeFeeFiat?: string;
+
+  /** Chain ID of the payment token. */
+  chainId?: Hex;
+
+  /** Total network fee in fiat currency, including the original and bridge transactions. */
+  networkFeeFiat?: string;
+
+  /** Address of the payment token that the transaction funds were sourced from. */
+  tokenAddress?: Hex;
+
+  /** Total cost of the transaction in fiat currency, including gas, fees, and the funds themselves. */
+  totalFiat?: string;
 };